Refactor to use Sequence object
The Sequence object lazily generates a bit sequence and caches
the bit sequence so it can be efficiently re-used. This approach
allows us to ensure that every cell is tested on the same set of
sequences for a particular generation (possibly important if we
want to try Lexicase or other diversity-maintaining schemes)
and should make our lives easier if we want to try out more
complicated tests with more than one type of sequence or
a different sequence-generating method.